v4 - automated selection, Necrite added
in this edition of the script, I have removed the cSelect function and replaced with one that detects the player's mining level. If the script detects it between selected level ranges (starting on a new ore when most of level benefits have been earned for optimal xp/hr) then it will select the correct option of available level ranges. This edition adds Necrite rocks.